chore: Update login package exports
