+++ packages/login/package.json
@@ -12,7 +12,7 @@
 		"check": "tsc",
 		"lint": "eslint .",
 		"format": "npx prettier --write .",
-		"test": "echo \"Error: no test specified\" && exit 0",
+		"test": "vitest run --coverage",
 		"test:crypto": "pnpm node --openssl-legacy-provider node_modules/vitest/vitest.mjs run --coverage"
 	},
 	"keywords": [],
