+++ packages/login/test/index.test.ts
@@ -1,0 +1,8 @@
+import type { NPSMessage } from "rusty-motors-shared";
+import { handleLoginData } from "../src/internal.js";
+import { describe, expect, it } from "vitest";
+import { MockNPSMessage } from "./MockNPSMessage.js";
+
+it("needs tests", () => {
+    expect(true).toBe(true);
+});
